ipc: refactor ipc handlers into ipc-sway.c
This is preparative work for the backward-compatible I3SOCK
implementation.
