+++ include/sway/ipc-json.h
@@ -4,6 +4,9 @@
 #include "sway/tree/container.h"
 #include "sway/input/input-manager.h"
 
+json_object *ipc_json_success(void);
+json_object *ipc_json_failure(const char *error);
+
 json_object *ipc_json_get_version(void);
 
 json_object *ipc_json_describe_disabled_output(struct sway_output *o);
